requirements.txt
gradio==4.32.0
faiss-cpu==1.8.0.post1
openai==1.55.3
httpx==0.27.0
llama-index-vector-stores-faiss==0.1.2
llama-index-readers-file==0.1.33
matplotlib==3.9.3
docx2txt==0.8
openpyxl==3.1.5
llama-index-core==0.10.67
uvicorn==0.30.6
fastapi==0.112.0
simplejson==3.19.3
pydantic==2.10.6
modelscope==1.18.0
langchain_community==0.2.16
transformers==4.44.2
llama_index.embeddings.huggingface==0.2.3
llama-index-embeddings-langchain==0.1.2
datasets==2.21.0
oss2==2.19.0
sortedcontainers==2.4.0
addict==2.4.0
chromadb>=0.4.18
pdfminer.six==20250324
jieba==0.42.1
rank_bm25==0.2.2